% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRegression-package.R
\docType{package}
\name{DMRegressionFreq-package}
\alias{DMRegressionFreq}
\alias{DMRegressionFreq-package}
\title{DMRegressionFreq: Fitting Dirichlet-Multinomial Regression Using Frequentist Methods}
\description{
This package provides fast options fitting the Dirichlet-Multinomial Regression using Frequentist methods via the help of TMB.
}
\author{
\strong{Maintainer}: Yi Jia Gu \email{yjgu@uwaterloo.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
